glr working, but inefficient
