+++ ast/src/lists/indices.rs
@@ -1,21 +1,5 @@
-use crate::Term;
-use rand::{rngs::StdRng, thread_rng, Rng, SeedableRng};
-
-#[derive(Clone, PartialEq, Debug, Hash)]
-pub struct IndexVector {
-    pub index: usize,
-    pub index_terms: Vec<Term>,
-}
-
-#[derive(Clone, PartialEq, Debug, Hash)]
-pub struct Indices(pub Vec<Index>);
-
-#[derive(Clone, PartialEq, Debug, Hash)]
-pub enum Index {
-    Const { index: Vec<i64> },
-    Random { n: i64, seed: i64 },
-    IndexAndTerm { index: Box<Index>, term: Term },
-}
+use crate::{Index, IndexVector, Indices};
+use rand::{rngs::StdRng, Rng, SeedableRng};
 
 impl Indices {
     pub fn get_indices_and_terms(&self, len_list: usize) -> Vec<IndexVector> {
