+++ src/testing/expect_tests/done/index_with_term.socool
