+++ src/testing/expect_tests/done/list_composed_with_op.socool
