+++ src/testing/expect_tests/done/list_fit_length.socool
