+++ src/testing/expect_tests/done/list_indices_with_op_2.socool
