+++ src/testing/expect_tests/done/list_named_with_indicies.socool
