+++ src/testing/expect_tests/done/list_with_indicies.socool
