+++ src/testing/expect_tests/done/list_with_random_and_index.socool
