+++ src/testing/expect_tests/done/simple_fun.socool
