+++ src/testing/expect_tests/mocks/list_named_indexed_fit_length.socool
@@ -10,7 +10,7 @@ list2 = {
 }
 
 main = {
-    list1 @ [0, 1] 
+    List list1 @ [0, 1] 
     > FitLength list2
 }
